+++ src/main/java/kitchenpos/dao/MenuDao.java
@@ -12,6 +12,4 @@ public interface MenuDao extends Repository<Menu, Long> {
     Optional<Menu> findById(Long id);
 
     List<Menu> findAll();
-
-    long countByIdIn(List<Long> ids);
 }
