+++ src/main/java/kitchenpos/dao/OrderLineItemDao.java
@@ -2,7 +2,7 @@ package kitchenpos.dao;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.order.OrderLineItem;
 import org.springframework.data.repository.Repository;
 
 public interface OrderLineItemDao extends Repository<OrderLineItem, Long> {
