+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -13,12 +13,12 @@ import kitchenpos.dao.fakedao.InMemoryMenuDao;
 import kitchenpos.dao.fakedao.InMemoryOrderDao;
 import kitchenpos.dao.fakedao.InMemoryOrderTableDao;
 import kitchenpos.domain.MenuFactory;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTableFactory;
 import kitchenpos.domain.TableGroup;
 import kitchenpos.domain.menugroup.MenuGroup;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderLineItem;
+import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.exception.CannotChangeEmptyException;
 import kitchenpos.exception.InvalidGuestNumberException;
 import kitchenpos.ui.request.OrderTableChangeEmptyRequest;
